refactor populateDreamscape method into modules
