+++ js/app/view.js
@@ -1,13 +1,11 @@
 var Auth = global.Auth
 var environment = global.environment
 var THREE = require('three')
-var geometryChooser = require('./../services/geometry-chooser')
-var getMatrixData = require('./../services/get-matrix-data')
 var moment = require('moment')
 var parseDreamString = require('./../services/parse-dream-string')
 var $ = require('jquery')
 require('bootstrap-jquery')
-var range = require('lodash.range')
+var generateMergedGeometriesFromDreams = require('./../services/generate-merged-geometries-from-dreams')
 
 // posts stuff to the dom
 var dreamsView = {
@@ -50,61 +48,19 @@ var dreamsView = {
 
   // takes dreams, decides where they're going to go
   populateDreamscape: function (dreams) {
+    var mergedGeometries = generateMergedGeometriesFromDreams(dreams)
 
-    allDreamsGeometry = new THREE.Geometry()
-    pickingGeometry = new THREE.Geometry()
-
-    dreams.forEach(function ( dream, i ) {
-
-      var color = new THREE.Color();
-      var quaternion = new THREE.Quaternion();
-      var matrix = new THREE.Matrix4();
-
-      var singleDreamGeom = geometryChooser(dream.sentiment)
-
-      var matrixData = getMatrixData(i)
-
-      quaternion.setFromEuler( matrixData.rotation, false );
-
-      // the matrix has the position, scale, and rotation of the object
-      matrix.compose( matrixData.position, quaternion, matrixData.scale );
-
-      var faceCountBeforeMerge = allDreamsGeometry.faces.length
-      allDreamsGeometry.merge(singleDreamGeom, matrix)
-      var faceCountAfterMerge = allDreamsGeometry.faces.length
-
-      var faceIndices = { // this is how we can find the faces for this particular geom, after it is merged into the single geom
-        low: faceCountBeforeMerge,
-        hi: faceCountAfterMerge
-      }
-
-      range(faceIndices.low, faceIndices.hi).forEach(function (faceIndex) { // this is to support dream.viewed feature yet to be added to the rails back end
-        allDreamsGeometry.faces[faceIndex].materialIndex = (dream.viewed) ? 1 : 0
-      })
-
-      // give the singleDreamGeom's vertices a color corresponding to the "id"
-      applyVertexColorsToGeometry( singleDreamGeom, color.setHex( i ) );
-
-      pickingGeometry.merge( singleDreamGeom, matrix );
-
-      environment.pickingData[ i ] = {
-        position: matrixData.position,
-        rotation: matrixData.rotation,
-        scale: matrixData.scale,
-        faceIndices: faceIndices
-      }
-    })
+    environment.pickingData = mergedGeometries.pickingData
 
     // allDreamsMesh is all of the dream objects merged together together
     var materials = [ environment.defaultMaterial, environment.viewedMaterial ]
-    environment.dreamsMesh = new THREE.Mesh( allDreamsGeometry, new THREE.MultiMaterial(materials) );
+    environment.dreamsMesh = new THREE.Mesh( mergedGeometries.allDreamsGeometry, new THREE.MultiMaterial(materials) );
     environment.addObjectToScene( environment.dreamsMesh );
 
-    environment.pickingMesh = new THREE.Mesh( pickingGeometry, environment.pickingMaterial )
+    environment.pickingMesh = new THREE.Mesh( mergedGeometries.pickingGeometry, environment.pickingMaterial )
     environment.pickingScene.add( environment.pickingMesh );
 
     environment.resetCameraPosition()
-
   },
 
   showInfoModal: function () {
@@ -159,15 +115,6 @@ function parseTagObjects (tagObjects) {
     tags.push(tagObjects[i].word)
   }
   return tags
-}
-
-function applyVertexColorsToGeometry (geometry, color) {
-  geometry.faces.forEach(function (face) {
-    var n = (face instanceof THREE.Face3) ? 3 : 4;
-    for (var j = 0; j < n; j++) {
-      face.vertexColors[j] = color;
-    }
-  })
 }
 
 module.exports = dreamsView
